requirements.txt
streamlit
requests
openai
python-dotenv
langchain
matplotlib
wordcloud
